requirements.txt
Flask
Flask-RESTful
Flask-Cors
pytest
pytest-flask
pandas
python-dotenv